add linter and fix issues
• move source from app/src/main/java/ → app/src/main/kotlin/
• add ktlint
• use android.lint
• use intellij for code style hints
• enforce linting through `pre-commit`-hook (installed on build)
