+++ app/build.gradle
@@ -31,6 +31,9 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+    sourceSets {
+        main.java.srcDirs += 'src/main/kotlin'
+    }
 }
 
 kotlin {
@@ -46,9 +49,14 @@ dependencies {
     implementation 'com.android.support.constraint:constraint-layout:1.1.0'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:0.22.5'
 
-    implementation 'com.contentful.java:java-sdk:+'
+    implementation 'com.contentful.java:java-sdk:10.1.0'
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+}
+
+afterEvaluate {
+    apply from: 'lint/teaLint.gradle'
+    tasks['preBuild'].dependsOn copyGitHooks
 }
