+++ app/src/androidTest/kotlin/com/contentful/tea/kotlin/ExampleInstrumentedTest.kt
@@ -1,0 +1,23 @@
+package com.contentful.tea.kotlin
+
+import android.support.test.InstrumentationRegistry
+import android.support.test.runner.AndroidJUnit4
+import junit.framework.Assert.assertEquals
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        assertEquals("com.contentful.tea.kotlin", appContext.packageName)
+    }
+}
