+++ app/src/test/java/com/contentful/tea/kotlin/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.contentful.tea.kotlin
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
